Added test file. Update csvreade.cpp

+++ csvreader.cpp
@@ -344,7 +344,7 @@ string computeFunction(string value, string **table, map<string, int> rowName, m
     case '*':
         return to_string(arg1_result_int * arg2_result_int);
     case '/':
-        return to_string(double(arg1_result_int / arg2_result_int));
+        return to_string(arg1_result_int / arg2_result_int);
     }
     return "_FUNCTION_ERROR_";
 }
